working on sme dashboard responsiveness
