+++ src/styles/dashboard/dashboard.css
@@ -1,5 +1,9 @@
 .dashboard{
-    padding: 40px;
+    margin: 0;
+    padding: 0;
+    padding-top: 40px;
+    margin-left: 80px;
+    margin-right: 50px;
     display: flex;
 }
 h1{
